# testdata/fiber/openapi.yaml
openapi: 3.0.3
info:
    title: My Awesome API
    description: |4-

        Copyright 2025 Ehab Terra. Licensed under the Apache License 2.0. See LICENSE and NOTICE.
    version: 1.0.0
    contact:
        name: Ehab
        url: https://ehabterra.github.io/
        email: ehabterra@hotmail.com
    license:
        name: ""
paths:
    /payment/payment/process:
        post:
            tags:
                - /payment
            operationId: github.com/ehabterra/swagen/testdata/fiber/payment.github.com/ehabterra/swagen/testdata/fiber/payment.ProcessPayment
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_gofiber_fiber_v2.Map'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
    /payment/stripe/pk:
        get:
            tags:
                - /payment
            operationId: github.com/ehabterra/swagen/testdata/fiber/payment.github.com/ehabterra/swagen/testdata/fiber/payment.GetStripePublicKey
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_gofiber_fiber_v2.Map'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
    /products/:
        get:
            tags:
                - /products
            operationId: github.com/ehabterra/swagen/testdata/fiber/products.github.com/ehabterra/swagen/testdata/fiber/products.ListProducts
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_products.Product'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
        post:
            tags:
                - /products
            operationId: github.com/ehabterra/swagen/testdata/fiber/products.github.com/ehabterra/swagen/testdata/fiber/products.CreateProduct
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_products.CreateProductRequest'
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_products.Product'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
    /products/{id}:
        get:
            tags:
                - /products
            operationId: github.com/ehabterra/swagen/testdata/fiber/products.github.com/ehabterra/swagen/testdata/fiber/products.GetProduct
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_gofiber_fiber_v2.Map'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
    /users/:
        get:
            tags:
                - /users
            operationId: github.com/ehabterra/swagen/testdata/fiber/users.github.com/ehabterra/swagen/testdata/fiber/users.ListUsers
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_users.User'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
        post:
            tags:
                - /users
            operationId: github.com/ehabterra/swagen/testdata/fiber/users.github.com/ehabterra/swagen/testdata/fiber/users.CreateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_users.CreateUserRequest'
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_users.User'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
    /users/{id}:
        get:
            tags:
                - /users
            operationId: github.com/ehabterra/swagen/testdata/fiber/users.github.com/ehabterra/swagen/testdata/fiber/users.GetUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_gofiber_fiber_v2.Map'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
        put:
            tags:
                - /users
            operationId: github.com/ehabterra/swagen/testdata/fiber/users.github.com/ehabterra/swagen/testdata/fiber/users.UpdateUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/github.com_ehabterra_swagen_testdata_fiber_users.UpdateUserRequest'
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_gofiber_fiber_v2.Map'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
        delete:
            tags:
                - /users
            operationId: github.com/ehabterra/swagen/testdata/fiber/users.github.com/ehabterra/swagen/testdata/fiber/users.DeleteUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/github.com_gofiber_fiber_v2.Map'
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
components:
    schemas:
        github.com_ehabterra_swagen_testdata_fiber_products.CreateProductRequest:
            type: object
            properties:
                name:
                    type: string
                price:
                    type: number
        github.com_ehabterra_swagen_testdata_fiber_products.Product:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                price:
                    type: number
        github.com_ehabterra_swagen_testdata_fiber_users.CreateUserRequest:
            type: object
            properties:
                email:
                    type: string
                name:
                    type: string
        github.com_ehabterra_swagen_testdata_fiber_users.UpdateUserRequest:
            type: object
            properties:
                email:
                    type: string
                name:
                    type: string
        github.com_ehabterra_swagen_testdata_fiber_users.User:
            type: object
            properties:
                email:
                    type: string
                id:
                    type: integer
                name:
                    type: string
        github.com_gofiber_fiber_v2.Map:
            type: object